+++ markdown-table/src/main/java/com/iffly/compose/markdown/table/MarkdownTable.kt
@@ -283,7 +283,7 @@ private fun TableTitle(
                 style = tableTheme.copyTextStyle,
                 modifier =
                     Modifier.clickable {
-                        actionHandler?.handleCopyClick(tableBlock)
+                        actionHandler.value?.handleCopyClick(tableBlock)
                     },
             )
         }