+++ markdown/src/main/java/com/iffly/compose/markdown/config/LocalActionHandlerProvider.kt
@@ -4,14 +4,17 @@ package com.iffly.compose.markdown.config
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.staticCompositionLocalOf
 import com.iffly.compose.markdown.ActionHandler
+import com.iffly.compose.markdown.ActionHandlerState
 
 internal val LocalActionHandlerProvider =
-    staticCompositionLocalOf<ActionHandler?> {
-        null
+    staticCompositionLocalOf<ActionHandlerState> {
+        mutableStateOf(null)
     }
 
 @Composable
 @ReadOnlyComposable
-fun currentActionHandler(): ActionHandler? = LocalActionHandlerProvider.current
+fun currentActionHandler(): ActionHandlerState = LocalActionHandlerProvider.current