+++ markdown/src/main/java/com/iffly/compose/markdown/core/renders/TextNodeStringBuilder.kt
@@ -1,7 +1,7 @@
 package com.iffly.compose.markdown.core.renders
 
 import androidx.compose.ui.text.AnnotatedString
-import com.iffly.compose.markdown.ActionHandler
+import com.iffly.compose.markdown.ActionHandlerState
 import com.iffly.compose.markdown.render.CompositeChildNodeStringBuilder
 import com.iffly.compose.markdown.render.IInlineNodeStringBuilder
 import com.iffly.compose.markdown.render.MarkdownInlineView
@@ -21,7 +21,7 @@ class TextNodeStringBuilder : IInlineNodeStringBuilder<Text> {
         node: Text,
         inlineContentMap: MutableMap<String, MarkdownInlineView>,
         markdownTheme: MarkdownTheme,
-        actionHandler: ActionHandler?,
+        actionHandler: ActionHandlerState?,
         indentLevel: Int,
         isShowNotSupported: Boolean,
         renderRegistry: RenderRegistry,