+++ markdown/src/main/java/com/iffly/compose/markdown/render/BaseNodeStringBuilder.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.ParagraphStyle
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.withStyle
-import com.iffly.compose.markdown.ActionHandler
+import com.iffly.compose.markdown.ActionHandlerState
 import com.iffly.compose.markdown.style.MarkdownTheme
 import com.iffly.compose.markdown.util.contentText
 import com.vladsch.flexmark.util.ast.Node
@@ -26,7 +26,7 @@ fun AnnotatedString.Builder.buildChildNodeAnnotatedString(
     inlineContentMap: MutableMap<String, MarkdownInlineView>,
     markdownTheme: MarkdownTheme,
     renderRegistry: RenderRegistry,
-    actionHandler: ActionHandler? = null,
+    actionHandler: ActionHandlerState? = null,
     isShowNotSupported: Boolean,
     nodeStringBuilderContext: NodeStringBuilderContext,
 ) {
@@ -106,7 +106,7 @@ open class CompositeChildNodeStringBuilder<T : Node> : IInlineNodeStringBuilder<
         node: T,
         inlineContentMap: MutableMap<String, MarkdownInlineView>,
         markdownTheme: MarkdownTheme,
-        actionHandler: ActionHandler?,
+        actionHandler: ActionHandlerState?,
         indentLevel: Int,
         isShowNotSupported: Boolean,
         renderRegistry: RenderRegistry,