+++ markdown/src/main/java/com/iffly/compose/markdown/render/BlockRendererInlineStringBuilder.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.text.style.BaselineShift
 import androidx.compose.ui.text.style.LineHeightStyle
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.sp
-import com.iffly.compose.markdown.ActionHandler
+import com.iffly.compose.markdown.ActionHandlerState
 import com.iffly.compose.markdown.style.MarkdownTheme
 import com.iffly.compose.markdown.util.StringExt
 import com.iffly.compose.markdown.widget.richtext.RichTextInlineContent
@@ -46,7 +46,7 @@ class BlockRendererInlineStringBuilder<T : Block>(
         node: T,
         inlineContentMap: MutableMap<String, MarkdownInlineView>,
         markdownTheme: MarkdownTheme,
-        actionHandler: ActionHandler?,
+        actionHandler: ActionHandlerState?,
         indentLevel: Int,
         isShowNotSupported: Boolean,
         renderRegistry: RenderRegistry,
@@ -88,7 +88,7 @@ class DocumentInlineStringBuilder : IInlineNodeStringBuilder<Document> {
         node: Document,
         inlineContentMap: MutableMap<String, MarkdownInlineView>,
         markdownTheme: MarkdownTheme,
-        actionHandler: ActionHandler?,
+        actionHandler: ActionHandlerState?,
         indentLevel: Int,
         isShowNotSupported: Boolean,
         renderRegistry: RenderRegistry,
@@ -158,7 +158,7 @@ private fun AnnotatedString.Builder.buildChildInlineNodeString(
     child: Node,
     inlineContentMap: MutableMap<String, MarkdownInlineView>,
     markdownTheme: MarkdownTheme,
-    actionHandler: ActionHandler?,
+    actionHandler: ActionHandlerState?,
     indentLevel: Int,
     isShowNotSupported: Boolean,
     renderRegistry: RenderRegistry,