+++ markdown/src/main/java/com/iffly/compose/markdown/render/MarkdownInlineText.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
-import com.iffly.compose.markdown.ActionHandler
+import com.iffly.compose.markdown.ActionHandlerState
 import com.iffly.compose.markdown.config.LocalNodeDataMap
 import com.iffly.compose.markdown.config.currentActionHandler
 import com.iffly.compose.markdown.config.currentRenderRegistry
@@ -211,7 +211,7 @@ fun markdownText(
     node: Node,
     markdownTheme: MarkdownTheme,
     renderRegistry: RenderRegistry,
-    actionHandler: ActionHandler? = null,
+    actionHandler: ActionHandlerState? = null,
     indentLevel: Int = 0,
     isShowNotSupported: Boolean,
     nodeStringBuilderContext: NodeStringBuilderContext,